+++ global.R
@@ -41,7 +41,13 @@ plotOutputInsideFluidRow <- function(plotOutputId) {
     fluidRow(
       column(12, align="center",
         tags$head(tags$style("#plot{height:80vh !important;}")),
-        plotOutput(plotOutputId),
+        plotOutput(plotOutputId,
+          dblclick=paste0(plotOutputId, "_dblclick"),
+          brush=brushOpts(
+            id=paste0(plotOutputId, "_brush"),
+            resetOnNew=TRUE
+          )
+        ),
         tags$style(type="text/css", "body { overflow-y: scroll; }")
       )
     ) %>% withSpinner(color="#0dc5c1")