+++ src/plots/plots.R
@@ -1,3 +1,15 @@
+coord_flip_wrapper <- function(..., default=FALSE) {
+  # Adds the argument "default" to the original ggplot's coord_flip.
+  # This gets rid of the annoying "Coordinate system already present. Adding
+  # new coordinate system, which will replace the existing one" warning which
+  # is otherwise issued when trying to modify the x and y limits in plots that
+  # use cood_flip
+  # Adapted from <https://github.com/tidyverse/ggplot2/issues/2799>
+  cf <- coord_flip(...)
+  cf$default <- default
+  return(cf)
+}
+
 plotTypesHierarchical <- list()
 plotTypesFlat <- list()