+++ src/plots/plots_vertical_profiles.R
@@ -14,8 +14,11 @@ doPlot.plotVerticalObsProfile <- function(p, plotRequest, plotData) {
     aes(x=level, y=obsvalue) +
     geom_line(colour="black") +
     geom_point(size=2, colour="black") +
-    coord_flip() +
+    coord_flip_wrapper(default=TRUE) +
     labs(x="Level", y=ylab)
+  # Mark that we will allow zooming in. This is not a feature of ggplot, but
+  # rather something for which we will handle the logic in the shiny server
+  attr(obplot, "allowZoom") <- TRUE
   return(obplot)
 }
 
@@ -28,8 +31,11 @@ doPlot.plotVerticalAnalysisProfile <- function(p, plotRequest, plotData) {
     geom_line() +
     geom_point(size=2) +
     scale_colour_manual(values=c("black", "blue", "red")) +
-    coord_flip() +
+    coord_flip_wrapper(default=TRUE) +
     labs(x="Level", y=ylab)
+  # Mark that we will allow zooming in. This is not a feature of ggplot, but
+  # rather something for which we will handle the logic in the shiny server
+  attr(obplot, "allowZoom") <- TRUE
   return(obplot)
 }
 
@@ -42,8 +48,11 @@ doPlot.plotVerticalBiasProfile <- function(p, plotRequest, plotData) {
     geom_line() +
     geom_point(size=2) +
     scale_colour_manual(values=c("black", "blue")) +
-    coord_flip() +
+    coord_flip_wrapper(default=TRUE) +
     labs(x="Level", y=ylab)
+  # Mark that we will allow zooming in. This is not a feature of ggplot, but
+  # rather something for which we will handle the logic in the shiny server
+  attr(obplot, "allowZoom") <- TRUE
   return(obplot)
 }